Ignore update
Updating the ignore file to include ring width files (.rwl), as well as any input files that are needed for the code and any processed files that are produced by the code.  This cuts down on using the git as a repository for data, and some data from some users may be proprietary.  we can link to a repository for example data if needed.